rm management of mysql datadir
Managing this dir causes an order of operations problem as the 'mysql'
user doesn't exist until after the mysql/mariadb packages have been
installed on the host. It is also unnecessary as the `mysql_datadir`
type used internal to the `mysql` mod will create this directory with
the appropriate ownership.